Issue #103: util.awaitableop: Move to sub-package to hide from Python 2.
At the moment, installing `pyhaystack` on Python 2 works, but it throws
an error (which it then ignores) when it tries to "compile" the `.py`
file to `.pyo`.

Since these files have no use in a Python 2 install, let's move them
into a sub-package which we can selectively import if and only if Python
3 is in use.
